+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `gantry follow [EXPERIMENT]` command for following the logs of a running experiment.
+
 ## [v0.17.2](https://github.com/allenai/beaker-gantry/releases/tag/v0.17.2) - 2023-07-21
 
 ### Fixed